[Glyphs Import] fix "TypeError: 'NoneType' object has no attribute '__getitem__'"
if `GlyphDict["layers"]` happens to be an empty list, `Layer` value is still None;
but then a subsequent `Layer['width']` fails with `TypeError` as one can't call
`__getitem__` on a `NoneType`.

so when a glyph's 'layers' does not contain a Layer whose id  matches the current
master id, we skip the current master id and continue the loop with the next master.